requirements.txt
streamlit>=1.28.0
google-adk>=0.1.0
google-genai>=0.3.0
requests>=2.31.0
aiohttp>=3.13.0
